Simplifying BufferProviders
