+++ provider/core/src/serde/de.rs
@@ -153,3 +153,40 @@ where
         }
     }
 }
+
+#[macro_export]
+macro_rules! impl_auto_deserializing {
+    ($buffer_provider: ty) => {
+        impl<M> ResourceProvider<M> for $buffer_provider
+        where
+            M: ResourceMarker,
+            // Actual bound:
+            //     for<'de> <M::Yokeable as Yokeable<'de>>::Output: serde::de::Deserialize<'de>,
+            // Necessary workaround bound (see `yoke::trait_hack` docs):
+            for<'de> yoke::trait_hack::YokeTraitHack<<M::Yokeable as yoke::Yokeable<'de>>::Output>:
+                serde::de::Deserialize<'de>,
+        {
+            fn load_resource(&self, req: &DataRequest) -> Result<DataResponse<M>, DataError> {
+                self.as_deserializing().load_resource(req)
+            }
+        }
+
+        impl<M> DynProvider<M> for $buffer_provider
+        where
+            M: DataMarker,
+            // Actual bound:
+            //     for<'de> <M::Yokeable as Yokeable<'de>>::Output: serde::de::Deserialize<'de>,
+            // Necessary workaround bound (see `yoke::trait_hack` docs):
+            for<'de> yoke::trait_hack::YokeTraitHack<<M::Yokeable as yoke::Yokeable<'de>>::Output>:
+                serde::de::Deserialize<'de>,
+        {
+            fn load_payload(
+                &self,
+                key: ResourceKey,
+                req: &DataRequest,
+            ) -> Result<DataResponse<M>, DataError> {
+                self.as_deserializing().load_payload(key, req)
+            }
+        }
+    };
+}
